implemented coroutines for background processing of database transactions. Moved the MovieRepository.kt file to the repository package.
