+++ Week6/MyFavoriteMovies/app/build.gradle
@@ -38,4 +38,6 @@ dependencies {
     testImplementation 'junit:junit:4.13'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.2.0"
 }
