+++ Week6/MyFavoriteMovies/app/src/main/java/com/raywenderlich/myfavoritemovies/ViewExtensions.kt
@@ -2,7 +2,7 @@ package com.raywenderlich.myfavoritemovies
 
 import android.view.View
 
-
+//this came from the solution, and I implemented it for the same reason as it was created.
 fun View.onClick(onClickAction: () -> Unit) {
     setOnClickListener { onClickAction() }
 }
