+++ Week6/MyFavoriteMovies/app/src/main/java/com/raywenderlich/myfavoritemovies/activities/DetailActivity.kt
@@ -4,13 +4,17 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
 import com.raywenderlich.myfavoritemovies.R
 import com.raywenderlich.myfavoritemovies.model.Movie
-import com.raywenderlich.myfavoritemovies.model.MovieRepository
+import com.raywenderlich.myfavoritemovies.repository.MovieRepository
 import kotlinx.android.synthetic.main.activity_detail.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 private const val MOVIE_KEY = "movie_id"
 
+//this came from the solution
 fun startDetailActivity(from: Context, movieId: Int?) =
     from.startActivity(Intent(from, DetailActivity::class.java).apply {
         putExtra(MOVIE_KEY, movieId)
@@ -24,9 +28,12 @@ class DetailActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_detail)
-
-        val movie = repository.getMovieById(movieId)
-        displayMovieDetails(movie)
+        lifecycleScope.launch(Dispatchers.IO) {
+            val movie = repository.getMovieById(movieId)
+            launch(Dispatchers.Main) {
+                displayMovieDetails(movie)
+            }
+        }
     }
 
     private fun displayMovieDetails(movie: Movie) {
