+++ Week6/MyFavoriteMovies/app/src/main/java/com/raywenderlich/myfavoritemovies/model/Movie.kt
@@ -1,11 +1,8 @@
 package com.raywenderlich.myfavoritemovies.model
 
-import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import kotlinx.android.parcel.Parcelize
 
-@Parcelize
 @Entity(tableName = "movies_table")
 class Movie(
     @PrimaryKey(autoGenerate = true) var id: Int,
@@ -14,4 +11,4 @@ class Movie(
     var summary: String,
     var genre: String,
     var poster: Int
-) : Parcelable
+)
